bff-requirements.txt
aiopulsar-py==0.1.0a0
fastapi==0.116.1
uvicorn[standard]==0.35.0
pulsar-client==3.7.0
pulsar-client[avro]==3.7.0
SQLAlchemy==2.0.41
requests==2.32.4
pydantic-settings==2.10.1
pydantic==2.10.0
PyMySQL==1.1.0